added functionality to test the internet connection on device and unittest for same
