+++ README.md
@@ -1,1 +1,7 @@
 # gitpy
+
+gitpy is a command line tool purely written in Python3 consumes GitHub developer's API.
+
+It provides the following features.
+
+1. Core : This package deals with authentication with GitHub using Authentication token.
