+++ tests/test_gitpy.py
@@ -2,7 +2,8 @@ import json
 import logging
 import unittest
 import os
-from gitpy import GitPy
+
+from core.gitpy import GitPy
 
 class TestGitPy(unittest.TestCase):
 
@@ -45,14 +46,30 @@ class TestGitPy(unittest.TestCase):
         self.logger.info('executing')
         username = self.configuration_data['username']
         token = self.configuration_data['token']
-        self.gitpy_object = GitPy(username=username,token=token+'nonce')
-        self.assertEqual(self.gitpy_object.authorization(),'Access Denied')
-        self.gitpy_object = GitPy(username=username+'nonce',token=token)
-        self.assertEqual(self.gitpy_object.authorization(),'Wrong Information')
         self.gitpy_object = GitPy(username=username,token=token)
-        self.assertEqual(self.gitpy_object.authorization(),'Authorization Successfull {}'.format(username))
-        self.logger.info('completed')
+        msg = self.gitpy_object.authorization()
+        if(self.gitpy_object.is_connected == False):
+            self.assertEqual(msg,'Please connect to Internet')
+        else:
+            self.gitpy_object = GitPy(username=username,token=token+'nonce')
+            self.assertEqual(self.gitpy_object.authorization(),'Access Denied : Wrong Token')
+            self.gitpy_object = GitPy(username=username+'nonce',token=token)
+            self.assertEqual(self.gitpy_object.authorization(),'Access Denied : Wrong Username')
+            self.gitpy_object = GitPy(username=username,token=token)
+            self.assertEqual(self.gitpy_object.authorization(),'Authorization Successfull {}'.format(username))
+            self.logger.info('completed')
 
+    def test_check_connectivity(self):
+        self.logger.info('executing')
+        username = self.configuration_data['username']
+        token = self.configuration_data['token']
+        self.gitpy_object = GitPy(username=username,token=token)
+        msg = self.gitpy_object.check_connectivity()
+        if(self.gitpy_object.is_connected):
+            self.assertEqual(msg,'Connected')
+        else:
+            self.assertEqual(smg,'Please connect to Internet')
+        self.logger.info('completed')
 
 if __name__ == '__main__':
     unittest.main()
